added images to attractions
