+++ js/attractionTypes.js
@@ -1,15 +1,12 @@
 "use strict";
 
 let db = require('./db-calls.js');
+let attractionTypesToDom = require("./attractionTypesToDom.js");
 
 
 
 let attractionTypes = {};
 
-function capitalLetter (string) {
-    return string.charAt(0).toUpperCase() + string.slice(1);
-    
-}
 
 
 db.fetchTypes()
@@ -17,20 +14,7 @@ db.fetchTypes()
        console.log("types", result);
         attractionTypes = result;
         return attractionTypes
-}).then(result => {
-    let fragment = document.createDocumentFragment();
-    ///////putting in the loop///////
-    result.forEach(function (type) {
-    let card = document.createElement("div");
-    card.className = "col-md-4";
-    let textArea = document.createElement("p");
-    textArea.innerHTML = capitalLetter(type.name);
-    card.appendChild(textArea);
-    fragment.appendChild(card);
-    });
-    //////////end the loop///////
-    return fragment
-})
+}).then(attractionTypesToDom)
 .then(result => {
     let divHolder = document.querySelector("#attractionType");
     divHolder.appendChild(result);
